+++ src/extensions/volume-tools/mask/index.ts
@@ -9,4 +9,7 @@
 
 export { VolumeMaskBehavior } from './behavior';
 export { MaskVolumeFromSource } from './transformers';
+export { MaskSelection } from './selection';
+export type { SelectionStore } from './selection';
+export { MaskSelectionColorThemeProvider, MaskSelectionColorThemeParams } from './theme';
 export type { ViewMask, MaskCreatorState, MaskSource, Point2D } from './types';